% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pjtemplate.R
\name{run_pjtemplate}
\alias{run_pjtemplate}
\alias{kickoff}
\title{Run create new project template}
\usage{
kickoff(path, ...)
}
\arguments{
\item{path}{new project path}

\item{...}{Variable-length argument.}
}
\value{
No return value.
}
\description{
Run create new project template
}
